workers/requirements.txt
redis>=5.0
requests>=2.31
aiohttp>=3.9
structlog>=24.1
Pillow>=10.3
pytesseract>=0.3.10
pandas>=2.2
pyyaml>=6.0
lxml>=5.2
faster-whisper>=1.0
pyttsx3>=2.90
urllib3>=2.2
httpx>=0.27
